# apps/backend/src/controller/payment.controller.ts
import {
  Controller,
  Get,
  Post,
  Put,
  Del,
  Body,
  Param,
  Query,
  Inject,
} from '@midwayjs/decorator';
import { Context } from '@midwayjs/koa';
import { Validate } from '@midwayjs/validate';
import { PaymentService } from '../service/payment.service';
import {
  CreatePaymentDto,
  UpdatePaymentDto,
  PaginationQuery,
  ApiResponse,
} from '../interface';

@Controller('/api/v1/payments')
export class PaymentController {
  @Inject()
  paymentService: PaymentService;

  @Inject()
  ctx: Context;

  /**
   * 创建支付记录
   */
  @Post('/')
  @Validate()
  async createPayment(
    @Body() createPaymentDto: CreatePaymentDto
  ): Promise<ApiResponse> {
    try {
      const rawBody = (this.ctx.request.body ?? {}) as Record<string, unknown>;
      const payerCustomerId =
        rawBody.payer_customer_id != null
          ? Number(rawBody.payer_customer_id)
          : null;

      const kitId = this.ctx.state?.kitId;
      if (!kitId) {
        return {
          success: false,
          message: '请选择套装',
          code: 400,
        };
      }

      // 验证发票属于当前 kit
      if (createPaymentDto.invoice_id) {
        const invoice = await this.paymentService.getInvoiceById(
          createPaymentDto.invoice_id
        );
        if (!invoice || invoice.kit_id !== kitId) {
          return {
            success: false,
            message: '发票不属于当前套装',
            code: 400,
          };
        }
      }

      // DTO 未声明 payer_customer_id，但 service 通过展开透传到 entity，结构兼容
      const createDto = {
        ...createPaymentDto,
        payer_customer_id: payerCustomerId,
      } as unknown as CreatePaymentDto;
      const payment = await this.paymentService.createPayment(createDto, kitId);
      return {
        success: true,
        data: payment,
        message: '支付记录创建成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '支付记录创建失败',
        code: 400,
      };
    }
  }

  /**
   * 获取支付记录列表
   */
  @Get('/')
  async getPayments(
    @Query()
    query: PaginationQuery & {
      invoiceId?: number;
      status?: string;
      viewAll?: string;
    }
  ): Promise<ApiResponse> {
    try {
      // 如果 viewAll=true，则不传 kitId（查看全部套账）
      const kitId =
        query.viewAll === 'true' ? undefined : this.ctx.state?.kitId;
      const result = await this.paymentService.getPayments(query, kitId);
      return {
        success: true,
        data: result,
        message: '获取支付记录列表成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取支付记录列表失败',
        code: 500,
      };
    }
  }

  /**
   * 根据ID获取支付记录详情
   */
  @Get('/:id')
  async getPaymentById(
    @Param('id') id: number,
    @Query('viewAll') viewAll?: string
  ): Promise<ApiResponse> {
    try {
      const kitId = viewAll === 'true' ? undefined : this.ctx.state?.kitId;
      const payment = await this.paymentService.getPaymentById(id, kitId);
      if (!payment) {
        return {
          success: false,
          message: '支付记录不存在',
          code: 404,
        };
      }
      return {
        success: true,
        data: payment,
        message: '获取支付记录详情成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取支付记录详情失败',
        code: 500,
      };
    }
  }

  /**
   * 更新支付记录信息
   */
  @Put('/:id')
  @Validate()
  async updatePayment(
    @Param('id') id: number,
    @Body() updatePaymentDto: UpdatePaymentDto
  ): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      if (!kitId) {
        return {
          success: false,
          message: '请选择套装',
          code: 400,
        };
      }

      // 若更换发票，验证新发票属于当前 kit
      if (updatePaymentDto.invoice_id) {
        const invoice = await this.paymentService.getInvoiceById(
          updatePaymentDto.invoice_id
        );
        if (!invoice || invoice.kit_id !== kitId) {
          return {
            success: false,
            message: '发票不属于当前套装',
            code: 400,
          };
        }
      }

      const rawBody = (this.ctx.request.body ?? {}) as Record<string, unknown>;
      const hasPayer = 'payer_customer_id' in rawBody;
      const payerCustomerId =
        rawBody.payer_customer_id != null
          ? Number(rawBody.payer_customer_id)
          : null;

      const mergedDto = hasPayer
        ? { ...updatePaymentDto, payer_customer_id: payerCustomerId }
        : updatePaymentDto;

      // DTO 未声明 payer_customer_id，但 service 通过展开透传到 entity，结构兼容
      const updateDto = mergedDto as unknown as UpdatePaymentDto;
      const payment = await this.paymentService.updatePayment(
        id,
        updateDto,
        kitId
      );
      if (!payment) {
        return {
          success: false,
          message: '支付记录不存在',
          code: 404,
        };
      }
      return {
        success: true,
        data: payment,
        message: '支付记录信息更新成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '支付记录信息更新失败',
        code: 400,
      };
    }
  }

  /**
   * 删除支付记录
   */
  @Del('/:id')
  async deletePayment(@Param('id') id: number): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      if (!kitId || !this.ctx.state?.user?.id)
        throw new Error('请登录并选择当前套账');
      const success = await this.paymentService.deletePayment(id, kitId);
      if (!success) {
        return {
          success: false,
          message: '支付记录不存在',
          code: 404,
        };
      }
      return {
        success: true,
        message: '支付记录删除成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '支付记录删除失败',
        code: 400,
      };
    }
  }

  /**
   * 根据发票ID获取支付记录
   */
  @Get('/invoice/:invoiceId')
  async getPaymentsByInvoiceId(
    @Param('invoiceId') invoiceId: number
  ): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      if (!kitId) {
        return {
          success: false,
          message: '请选择套装',
          code: 400,
        };
      }

      const invoice = await this.paymentService.getInvoiceById(invoiceId);
      if (!invoice || invoice.kit_id !== kitId) {
        return {
          success: false,
          message: '发票不属于当前套装',
          code: 400,
        };
      }

      const payments = await this.paymentService.getPaymentsByInvoiceId(
        invoiceId
      );
      return {
        success: true,
        data: payments,
        message: '获取支付记录成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取支付记录失败',
        code: 500,
      };
    }
  }

  /**
   * 获取支付统计信息
   */
  @Get('/stats/overview')
  async getPaymentStats(
    @Query() query: { viewAll?: string }
  ): Promise<ApiResponse> {
    try {
      // 如果 viewAll=true，则不传 kitId（查看全部套账）
      const kitId =
        query.viewAll === 'true' ? undefined : this.ctx.state?.kitId;
      const stats = await this.paymentService.getPaymentStats(undefined, kitId);
      return {
        success: true,
        data: stats,
        message: '获取支付统计成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取支付统计失败',
        code: 500,
      };
    }
  }
}