# docs/api/backend_service/api_spec.yaml
openapi: 3.0.3
info:
  title: ProContractLedger API (合同台账管理系统 API)
  version: v1.0.0
  description: |-
    用于管理ProContractLedger（合同台账管理系统）的客户、合同、发票和付款的API。
    此API基于提供的数据库结构和核心SQL查询进行设计。
  contact:
    name: API技术支持
    email: support@example.com
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: 开发服务器
  - url: https://api.procontractledger.com/v1
    description: 生产服务器
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT认证令牌。在请求头中添加：Authorization: Bearer <token>"
  schemas:
...
    PaymentBase:
      type: object
      description: 基础付款信息。对应 `payments` 表。
      properties:
        payment_id:
          type: integer
          format: int32
          description: 付款的唯一标识符。
          readOnly: true
        invoice_id:
          type: integer
          format: int32
          description: 关联发票的标识符。
        amount:
          type: string # 十进制数使用字符串
          format: decimal
          description: 付款金额。
        payment_date:
          type: string
          format: date-time
          description: 付款日期。
        payment_method:
          type: string
          maxLength: 50
          description: 付款方式 (例如：银行转账, 支付宝)。
        reference_number:
          type: string
          maxLength: 50
          nullable: true
          description: 付款参考号 (例如：交易ID)。
        notes:
          type: string
          format: text
          nullable: true
          description: 关于付款的附加备注。
        created_at:
          type: string
          format: date-time
          description: 付款记录创建时间戳。
          readOnly: true
      required:
        - invoice_id
        - amount
        - payment_date
        - payment_method
    PaymentCreateRequest:
      type: object
      description: 创建新付款记录的请求体。
      properties:
        invoice_id:
          type: integer
          format: int32
          description: 关联发票的标识符。
          example: 1
        amount:
          type: string
          format: decimal
          description: 付款金额。
          example: "2500.00"
        payment_date:
          type: string
          format: date
          description: 付款日期。
          example: "2023-04-01"
        payment_method:
          type: string
          maxLength: 50
          description: 付款方式 (例如：银行转账, 支付宝)。
          example: "银行转账"
        reference_number:
          type: string
          maxLength: 50
          nullable: true
          description: 付款参考号 (例如：交易ID)。
          example: "BANK-TXN-12345"
        notes:
          type: string
          format: text
          nullable: true
          description: 关于付款的附加备注。
          example: "第一期款项已到账"
      required:
        - invoice_id
        - amount
        - payment_date
        - payment_method
    PaymentUpdateRequest:
      type: object
      description: 更新付款记录的请求体。所有字段都是可选的，只提供需要更新的字段。
      properties:
        amount:
          type: string
          format: decimal
          description: 付款金额。
          example: "2600.00"
        payment_date:
          type: string
          format: date
          description: 付款日期。
          example: "2023-04-05"
        payment_method:
          type: string
          maxLength: 50
          description: 付款方式 (例如：银行转账, 支付宝)。
          example: "支付宝"
        reference_number:
          type: string
          maxLength: 50
          nullable: true
          description: 付款参考号 (例如：交易ID)。
          example: "ALIPAY-TXN-67890"
        notes:
          type: string
          format: text
          nullable: true
          description: 关于付款的附加备注。
          example: "金额调整，已收到尾款"
    PaymentResponse:
      type: object
      description: 付款记录响应体。
      allOf:
        - $ref: '#/components/schemas/PaymentBase'
      properties:
        payment_id:
          readOnly: true
        created_at:
          readOnly: true
        invoice_number:
          type: string
          description: 关联发票的编号。
          readOnly: true
        invoice_amount:
          type: string
          format: decimal
          description: 关联发票的金额。
          readOnly: true
        invoice_status:
          type: string
          description: 关联发票的状态。
          readOnly: true
        contract_id:
          type: integer
          format: int32
          description: 关联合同的ID。
          readOnly: true
        contract_number:
          type: string
          description: 关联合同的编号。
          readOnly: true
        contract_name:
          type: string
          description: 关联合同的名称。
          readOnly: true
        customer_id:
          type: integer
          format: int32
          description: 关联客户的ID。
          readOnly: true
        customer_name:
          type: string
          description: 关联客户的名称。
          readOnly: true
...
security:
  - BearerAuth: []
paths:
...
  /payments:
    get:
      summary: 获取所有付款记录
      tags:
        - Payments
      parameters:
        - in: query
          name: invoice_id
          schema:
            type: integer
            format: int32
          description: 发票ID
        - in: query
          name: payment_method
          schema:
            type: string
          description: 付款方式模糊搜索
        - in: query
          name: payment_date_from
          schema:
            type: string
            format: date
          description: 付款日期从 (YYYY-MM-DD)
        - in: query
          name: payment_date_to
          schema:
            type: string
            format: date
          description: 付款日期到 (YYYY-MM-DD)
        - in: query
          name: page
          schema:
            type: integer
            format: int32
            minimum: 1
          description: 页码，默认为1
        - in: query
          name: pageSize
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
          description: 每页数量，默认为10
      responses:
        200:
          description: 成功获取付款记录列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentResponse'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pageSize:
                        type: integer
        500:
          description: 获取付款记录列表失败
    post:
      summary: 创建付款记录
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        201:
          description: 付款记录创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "付款记录创建成功"
                  payment_id:
                    type: integer
                    format: int32
        400:
          description: 缺少必要的付款信息
        404:
          description: 关联的发票不存在
        500:
          description: 创建付款记录失败
  /payments/{paymentId}:
    get:
      summary: 获取单个付款记录
      tags:
        - Payments
      parameters:
        - in: path
          name: paymentId
          schema:
            type: integer
            format: int32
          required: true
          description: 付款ID
      responses:
        200:
          description: 成功获取付款记录信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        404:
          description: 付款记录不存在
        500:
          description: 获取付款记录详情失败
    put:
      summary: 更新付款记录
      tags:
        - Payments
      parameters:
        - in: path
          name: paymentId
          schema:
            type: integer
            format: int32
          required: true
          description: 付款ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentUpdateRequest'
      responses:
        200:
          description: 付款记录更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "付款记录更新成功"
        400:
          description: 付款记录未做任何更改
        404:
          description: 付款记录不存在
        500:
          description: 更新付款记录失败
    delete:
      summary: 删除付款记录
      tags:
        - Payments
      parameters:
        - in: path
          name: paymentId
          schema:
            type: integer
            format: int32
          required: true
          description: 付款ID
      responses:
        200:
          description: 付款记录删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "付款记录删除成功"
        404:
          description: 付款记录不存在
        500:
          description: 删除付款记录失败
...